Add a Properties test suite
The suite currently only contains one test, that checks whether a value
passed to Set is retrieved back when queried to Get.
The key and value are defined as constants.
The test suite is declared in the same package, not a distinct test one
because I was not able to make this work.