Add tests for wrapped lines
One test ensures that properties.Load proceses wrapped lines correctly
(discards the backslash and leading whitespace on the continuation line
and append the rest to the initial line); the other ensures that the
method does not accept trailiing backslashes without a continuation line
(i.e. a backslash on the last line of the file).