Change whitespace trimming strategy
Whitespace is now trimmed around the key and value in Properties.Load
using strings.Trim, with an additional string parameter containing the
space and tab characters only. This ensures consistent behavior with the
leading trimming used for wrapped lines, and restricts the list of
characters that should be trimmed (for example, line break characters
must not be considered "whitespace" in this context. In fact, they
should not even appear, and so they might not be the best example...)