Extract functions for future evolutions
These functions are not yet implemented; they only encapsulate
statements or expressions that must evolve to advance the state of the
project. The first of these functions abstracts away the splitting of a
logical line into the key and value (an operation that is not as simple
as "split around the first separator", as the latter must not be
escaped). The other two are used when writing out the properties, and
escape special characters in the key and value, respectively.