Inline internal function splitLine
This is the first step to try and fix the issue of the method Load
potentially reporting an incorrect line number. It occurs because Load
only retains the number of the last continuation line of a property
definition, so the discrepancy manifests when the physical line where
the error occurs is either the first line or a continuation line that is
not the last one.
Fixing this issue is not possible in the current structure of the code,
since the line number to report is for a physical line, whereas the
function splitLine is only given logical lines.