Handle comments using a state variable
Like the other particular handlings required in Properties.Load,
detection and processing of comment lines is controlled with a boolean
variable, instead of reusing the reader in the big if/else block.
This introduces more state, and probably makes execution times a bit
worse, but the reader is now entirely decoupled from the character
processing block.