Use an instance of the Load state structure
The instance is created in Load, with initial values explicitly set for
the fields where the default value is incorrect. It is passed to the
internal function processByte, instead of the seven pointers to th local
state variables. This simplifies the function's prototype greatly, and
will simplify evolution (by removing the need to declare a new function
parameter) as well.