Allow escaped separators also in property values
The boolean parameter to the unescape function indicating whether the
parser is in a key or not is removed; instead, an escaped separator is
processed independently of context (that is, duly unescaped).
Having escaped separators in the value is not technically necessary,
but helps convey to human readers the fact that this equals sign is not
intended as a separator.